Update Irredentism.txt
Add back missing estonian culture for Greater Finland. In the old PDM mod it have a "embrace_estonians" decisions for both SCA and FIN.
Estonia have a much, much closer relationship in history than the "ugrian" in game, 
https://en.wikipedia.org/wiki/Estonia%E2%80%93Finland_relations
Unification proposals
In Finland, the idea of a unified Greater Finland gained popularity and influence rapidly in 1917, but lost support after World War II and the Continuation War.
In Estonia of 1917, 1918 and 1940, the president Konstantin Päts wished for a union of finnic countries in his political testament, writing an outline of a unification plan. He used the term "soomesugu" in Estonian to point to the people and countries of Finland and Estonia when talking about common border law.[3][4]